add boxing, enum, data, BigDecimal

+++ src/com_algaworks_banco/modelo/pagamento/Boleto.java
@@ -1,20 +1,22 @@
 package com_algaworks_banco.modelo.pagamento;
 
+import java.math.BigDecimal;
+
 import com_algaworks_banco.modelo.Pessoa;
 
 public class Boleto implements DocumentoPagavel, DocumentoEstornavel {
 
     private Pessoa beneficiario;
-    private double valor;
+    private BigDecimal valor;
     private boolean pago;
 
-    public Boleto(Pessoa beneficiario, double valor) {
+    public Boleto(Pessoa beneficiario, BigDecimal valor) {
         this.beneficiario = beneficiario;
         this.valor = valor;
     }
 
     @Override
-    public double getValorTotal() {
+    public BigDecimal getValorTotal() {
         return valor;
     }
